pyproject.toml
[tool.poetry]
name = "aidebates"
version = "0.0.1"
description = "AI Debates for AmpCamp 2023"
authors = ["Denis Volk <the.denis.volk@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = []


[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
numpy = "^1.24.2"
retry = "^0.9.2"
python-dateutil = "^2.8.2"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
pydantic = "<2"
pre-commit = "^3.3.3"
pylint = "^2.17.2"
flake8 = "^6.0.0"
mypy = "^1.2.0"
isort = "^5.12.0"
black = "^23.3.0"
types-requests = "^2.28.11.17"
types-retry = "^0.9.9.3"
types-pyyaml = "^6.0.12.9"
types-python-dateutil = "^2.8.19.14"
jupyterlab = "^4.0.2"
ipykernel = "^6.22.0"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
ipympl = "^0.9.3"
plotly = "^5.14.1"
tqdm = "^4.65.0"
pylint-pydantic = "^0.2.4"
pydeps = "^1.12.17"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line_length = 119
target_version = ['py310']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.github
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | notebooks
    | \.ipynb_checkpoints
    | __pycache__
    | data
    | logs
    | _build
    | buck-out
    | build
    | dist
    | snapshots
  )/
)
'''